feat: drop GopherPanel into the install wizard, mount /settings/gopher in setup mode
Install wizard's Gopher section used to be two cramped inputs (URL +
key) tucked into the Optional card — no subdomain field, no inline
verification, no bootstrap visibility. Operators saved blind, finished
the wizard, and discovered failures only on the dashboard.

Replaced with the same GopherPanel the Infrastructure → Gopher tunnels
page renders. Now:

  - Toggle: 'Enable Gopher tunnels' (off by default). Off = skip
    Gopher entirely, finish the wizard normally.
  - On: GopherPanel appears (the existing component, lifted out of
    GopherTunnels.tsx into components/ui/GopherPanel.tsx). Operator
    pastes URL + key + subdomain, clicks Save, watches the
    SelfBootstrapModal walk through registering → installing →
    waiting_connect → creating_tunnel → active.
  - Next button blocks until cloud tunnel reaches active OR the
    operator flips the toggle back off. Surfaces a one-line hint
    explaining the gate.

Backend wiring:
  - Setup-mode router now initialises DB + AuthService + selftunnel
    so /api/settings/gopher/* endpoints (GET/PUT + self-bootstrap
    status/start) work pre-configuration. Same handler code the
    configured-mode router uses — no duplicate logic.
  - main.go's setup-mode branch opens the DB (User + GopherSettings
    models) and wires selftunnel.Service. Best-effort: DB init
    failure logs + falls back to the original no-gopher wizard so
    the install wizard never hard-fails on the new dependency.
  - Removed gopher_api_url / gopher_api_key from the setup save
    payload — GopherPanel saves directly via PUT /api/settings/gopher
    before the wizard advances.

Shared component reuse + no behaviour drift between the two surfaces.
